fix:Not adm users cannot delete invites that not own

+++ src/services/invites/inviteDelete.service.ts
@@ -1,6 +1,16 @@
+import { AppError } from "../../errors/errors";
 import { inviteRepository } from "../../repositories/inviteRepository";
 
-export const deleteInviteService = async (id: string): Promise<void> => {
-  const invite = await inviteRepository.findOneBy({ id: id });
+export const deleteInviteService = async (id: string,user: {id: string,type: string,isAdm: boolean}): Promise<void> => {
+  const invite = await inviteRepository.createQueryBuilder("invites")
+  .where("invites.id = :id_invite",{id_invite:id})
+  .leftJoinAndSelect("invites.userIdReceive","receiver")
+  .leftJoinAndSelect("invites.userIdSend","sender")
+  .getOne()
+
+  if(!(user.id==invite?.userIdReceive.id || user.id == invite?.userIdSend.id || user.isAdm)){
+    throw new AppError (401, 'You cannot delete an invitation that you do not own.');
+  }
+
   await inviteRepository.remove(invite!);
 };
